yearpicker as a component + added comments
